pyproject.toml
[project]
name = "launchchit"
version = "0.1.0"
description = "Platform for African developers to publish apps for community testing, feedback, and voting."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "aiosqlite>=0.20",
    "pydantic-settings>=2.6",
    "alembic>=1.14",
    "python-dotenv>=1.0",
    "python-jose[cryptography]>=3.3",
    "aiosmtplib>=3.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
